add routing to the landing page
