--- src/app/app.component.html
@@ -1,1 +1,0 @@
-<gd-map-view class="root map-view"></gd-map-view>
