+++ src/app/app.component.ts
@@ -1,9 +1,12 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'gd-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
+  selector: 'gd-root', 
+  template: `
+  <div class="root">
+    <router-outlet></router-outlet>
+  </div>
+`
 })
 
 export class AppComponent {
