+++ src/app/gd-routing.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { DashboardComponent } from './pages/dashboard/dashboard.component';
+
+import { MapViewComponent } from './pages/map-view/map-view.component';
+
+const routes: Routes = [
+  { path: '', redirectTo: '/dashboard', pathMatch: 'full' },
+  { path: 'dashboard', component: DashboardComponent },
+  { path: 'mapview', component: MapViewComponent },
+  { path: 'mapview/:name', component: MapViewComponent },
+];
+
+@NgModule({
+  imports: [ RouterModule.forRoot(routes) ],
+  exports: [ RouterModule ]
+})
+export class GdRoutingModule { }
