+++ src/app/pages/dashboard/dashboard.component.scss
@@ -1,0 +1,57 @@
+.dashboard {
+    justify-content: center;
+    align-items: center;
+    display: flex;
+    > .cards-selector {
+        margin: 9rem 5rem;   
+    }
+}
+
+.cards-selector {
+    margin: 9rem 5rem;
+
+    
+    border: 1px solid #F0F0F0;
+    display: inline-flex;
+    flex-flow: column nowrap;
+    align-items: center;
+    padding: 2rem;
+    
+}
+
+.card-tiles {
+    display: flex;
+    flex-flow: row nowrap;
+
+    > .tile {
+        color: inherit;
+        text-decoration: none;
+        
+        margin: 1rem;
+        display: flex;
+        flex-direction: column;
+        height: 24rem;
+        width: 24rem;
+        align-items: center;
+        transition: opacity 1s;
+        border: 1px solid rgba(150,150,150,0.3);
+        justify-content: space-between;
+
+        &:hover {
+            border-color: rgba(150,150,150,1);
+            box-shadow: 0 1px 2px 0 rgba(0,0,0,0.1), 0 4px 8px 0 rgba(0,0,0,0.2);
+            transition-duration: 200ms;
+            transition-property: transform, border, box-shadow, margin, opacity, width;
+        }
+
+        > .title {
+            padding-top: 1.6rem;
+            line-height: 2.4rem;
+            text-align: center;
+        }
+
+        > .thumb {
+            padding-bottom: 1.6rem;
+        }
+    }
+}
