+++ src/app/pages/map-view/sidebar-section/sidebar-section.component.ts
@@ -1,44 +1,18 @@
 import { Component, ContentChild, ElementRef, EventEmitter, Input, OnInit, OnChanges, Output, Renderer2, ViewChild } from '@angular/core';
 import { LayerService } from '../../../services/layer.service';
-import * as chroma from 'chroma-js';
-import { LeafletDirective, LeafletDirectiveWrapper } from '@asymmetrik/ngx-leaflet';
 import * as L from 'leaflet';
 import 'leaflet-draw';
 
-export interface LayerCard {
-    info: {
-        name: string;
-        title: string;
-        presets: [{
-            weights: number[],
-            text: string,
-        }];
-        optional?: string[];
-        prtext: string[];
-        range: number;
-        palettes?: string[];
-    };
-    weights: number[];
-    show: boolean;
-    opacity: number;
-    mask?: string;
-    params: string;
-    wmsServer: string;
-    sumServer?: string;
-    palette: string[] | string;
-    summary?: {
-        layers: number[],
-        total: number
-    };
-}
+import { LayerCard } from '../../../layer-card';
 
 @Component({
     selector: 'gd-sidebar-section',
     templateUrl: './sidebar-section.component.html',
-    providers: [LayerService, LeafletDirective]
+    providers: [LayerService]
 })
-export class SidebarSectionComponent implements OnChanges {
+export class SidebarSectionComponent implements OnInit, OnChanges {
     @Input() map: L.Map;
+    @Input() cards: LayerCard[];
     @Input() mask = '';
     @Output() onSummaryChange = new EventEmitter<object>();
 
@@ -50,99 +24,9 @@ export class SidebarSectionComponent implements OnChanges {
     colorNumber = 10;
     colorPalette: string[];
 
-    paramsLM: string[] = ['philly_bars', 'philly_grocery_stores', 'philly_rail_stops'];
-    paramsCHT = 'ImperviousSurfaces_Barren Lands_Open Water,DevelopedLand,Wetlands,ForestedLands,Non-workingProtectedOrPublicLands,PublicallyOwnedWorkingLands,PrivatelyOwnedWorkingLandsWithEasements,FarmlandWithoutPrimeAgriculturalSoils,FarmlandOrForestedLandsWithPrimeAgriculturalSoils';
-
     layersMap: Map<string, L.Layer> = new Map();
 
     layers: L.Layer[] = [];
-
-    lmCard: LayerCard = {
-        info: {
-            name: 'lm',
-            title: 'Philly Location Modeling',
-            presets: [
-                {
-                    weights: [3, -3, 0],
-                    text: 'NEAR Bars, BUT NOT Grocery Stores'
-                }, {
-                    weights: [0, -3, 3],
-                    text: 'NEAR Grocery Stores, BUT NOT Bars'
-                }, {
-                    weights: [0, -3, 3],
-                    text: 'NEAR Rail Stops, BUT NOT Grocery Stores'
-                }, {
-                    weights: [3, 0, 0],
-                    text: 'NEAR Bars'
-                }],
-            prtext: this.paramsLM.map(el => {
-                return el.slice(7);
-            }),
-            range: 4,
-        },
-        params: this.paramsLM.join(),
-        weights: [3, -3, 0],
-        show: true,
-        opacity: 0.6,
-        wmsServer: 'https://geotrellis.io/gt/weighted-overlay/wms',
-        palette: chroma.scale(this.colorLM).mode('lab').domain([0, 0.5, 0.6, 1]).colors(10)
-
-    };
-
-    chattaCard: LayerCard = {
-        info: {
-            name: 'chatta',
-            title: 'Chattanooga Agriculture & Forestry Value Model',
-            presets: [
-                {
-                    weights: [-5, -4, -2, 1, -1, 2, 4, 3, 5],
-                    text: 'Preset1'
-                }, {
-                    weights: [-5, 4, -2, 1, 1, 2, 4, 3, 5],
-                    text: 'Preset2'
-                }
-            ],
-            optional: ['summary'],
-            prtext: [
-                'Impervious_Surfaces_Barren_Lands_Open_Water',
-                'Developed_Land',
-                'Wetlands',
-                'Forested_Lands',
-                'Non-working_Protected_Or_Public_Lands',
-                'Publically_Owned_Working_Lands',
-                'Privately_Owned_Working_Lands_With_Easements',
-                'Farmland_Without_Prime_Agricultural_Soils',
-                'Farmland_Or_Forested_Lands_With_Prime_Agricultural_Soils'
-            ],
-            range: 6,
-            palettes: [
-                'yellow-to-red-heatmap',
-                'blue-to-yellow-to-red-heatmap',
-                'dark-red-to-yellow-heatmap',
-                'blue-to-red',
-                'green-to-red-orange',
-                'muted-terrain-qualitative',
-                'green-to-orange',
-                'light-to-dark-green',
-                'light-to-dark-sunset',
-                'purple-to-dark-purple-to-white-heatmap',
-                'blue-to-orange',
-                'bold-land-use-qualitative'
-            ],
-        },
-        params: this.paramsCHT,
-        weights: [-5, -4, -2, 1, -1, 2, 4, 3, 5],
-        show: true,
-        opacity: 0.6,
-        wmsServer: 'http://demo.geotrellis.com/chatta/gt/wo',
-        sumServer: 'http://demo.geotrellis.com/chatta/gt/sum',
-        mask: '',
-        palette: 'yellow-to-red-heatmap',
-        summary: null,
-    };
-
-    cards: LayerCard[] = [this.lmCard, this.chattaCard];
-    mapWrapper: LeafletDirectiveWrapper;
 
     getView(): void {
         this.map.eachLayer(el => {
@@ -280,10 +164,11 @@ export class SidebarSectionComponent implements OnChanges {
         private _el: ElementRef,
         private _rd: Renderer2,
         private _layerService: LayerService,
-        private ld: LeafletDirective
-
     ) {
-        this.mapWrapper = new LeafletDirectiveWrapper(this.ld);
+    }
+    ngOnInit() {
+        // for debugging
+        // console.log(this.cards, 'init');
     }
 
     ngOnChanges(changes) {
